refactor: 주문 계층별 패키징
