+++ src/main/java/com/main_39/Spring/order/repository/OrderRepository.java
@@ -1,5 +1,6 @@
-package com.main_39.Spring.order;
+package com.main_39.Spring.order.repository;
 
+import com.main_39.Spring.order.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
