+++ src/main/java/com/main_39/Spring/order/service/OrderService.java
@@ -1,15 +1,13 @@
-package com.main_39.Spring.order;
+package com.main_39.Spring.order.service;
 
 import com.main_39.Spring.exception.BusinessLogicException;
 import com.main_39.Spring.exception.ExceptionCode;
 import com.main_39.Spring.member.entity.Kakao;
 import com.main_39.Spring.member.repository.KakaoRepository;
-import com.main_39.Spring.menu.Menu;
-import com.main_39.Spring.menu.MenuRepository;
+import com.main_39.Spring.order.entity.Order;
+import com.main_39.Spring.order.repository.OrderRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
